Refactor config declarations into own folder
