+++ src/config/weatherConfig.ts
