+++ src/config/weatherConfigData.ts
